deletion-authorization

+++ app/api/deleteRecipe/route.js
@@ -1,32 +1,48 @@
 import { prisma } from "@/lib/prisma";
 
 export const POST = async (req, res) => {
-  const { recipeId } = await req.json();
+  const { recipeId, userId } = await req.json();
 
   try {
-    let recipe = await prisma.recipe.delete({
+    const recipe = await prisma.recipe.findUnique({
       where: {
         id: recipeId,
       },
     });
-    try {
-      let fav = await prisma.favourite.delete({
+
+    if (recipe && recipe.user_id === userId) {
+      await prisma.favourite.deleteMany({
         where: {
           recipe_id: recipeId,
         },
       });
-    } catch (err) {
-      console.log(err);
+
+      await prisma.recipe.delete({
+        where: {
+          id: recipeId,
+        },
+      });
+
+      return new Response(
+        JSON.stringify({
+          message: "Delete Successful!",
+        }),
+        {
+          status: 200,
+          headers: { "Content-Type": "application/json" },
+        }
+      );
+    } else {
+      return new Response(
+        JSON.stringify({
+          message: "Unauthorized!",
+        }),
+        {
+          status: 403,
+          headers: { "Content-Type": "application/json" },
+        }
+      );
     }
-    return new Response(
-      JSON.stringify({
-        message: "Delete Succesfull!",
-      }),
-      {
-        status: 200,
-        headers: { "Content-Type": "application/json" },
-      }
-    );
   } catch (err) {
     console.log(err);
     return new Response(JSON.stringify({ message: "Fetching User Failed" }), {
